+++ __tests__/frontend/scenes.tests.js
@@ -1,13 +1,13 @@
 import React from 'react';
-import AccountHome from '../src/scenes/AccountHome';
-import AccountSettings from '../src/scenes/AccountSettings';
-import Camera from '../src/scenes/Camera';
-import CameraGallery from '../src/scenes/CameraGallery';
-import FoodSelector from '../src/scenes/FoodSelector';
-import History from '../src/scenes/History';
-import LoggedInLanding from '../src/scenes/LoggedInLanding';
-import NutritionHistory from '../src/scenes/NutritionHistory';
-import Tab from '../src/scenes/NutritionHistory';
+import AccountHome from '../../src/scenes/AccountHome';
+import AccountSettings from '../../src/scenes/AccountSettings';
+import Camera from '../../src/scenes/Camera';
+import CameraGallery from '../../src/scenes/CameraGallery';
+import FoodSelector from '../../src/scenes/FoodSelector';
+import History from '../../src/scenes/History';
+import LoggedInLanding from '../../src/scenes/LoggedInLanding';
+import NutritionHistory from '../../src/scenes/NutritionHistory';
+import Tab from '../../src/scenes/NutritionHistory';
 
 import renderer from 'react-test-renderer';
 import { createStore } from 'redux';
